Remove FeedLoader protocol as we don't need it anymore - we're composing the types with universal abstractions provided by the Combine framework
