+++ EssentialApp/EssentialApp/CombineHelpers.swift
@@ -55,7 +55,7 @@ private extension FeedImageDataCache {
     }
 }
 
-public extension FeedLoader {
+public extension LocalFeedLoader {
     typealias Publisher = AnyPublisher<[FeedImage], Error>
     
     func loadPublisher() -> AnyPublisher<[FeedImage], Error> {
