--- EssentialApp/EssentialAppTests/Helpers/FeedLoaderStub.swift
@@ -1,20 +1,0 @@
-//
-//  FeedLoaderStub.swift
-//  EssentialAppTests
-//
-//  Created by Alexandre Gravelle on 2021-11-07.
-//
-
-import EssentialFeed
-
-class FeedLoaderStub: FeedLoader {
-    private let result: FeedLoader.Result
-    
-    init(result: FeedLoader.Result) {
-        self.result = result
-    }
-    
-    func load(completion: @escaping (FeedLoader.Result) -> Void) {
-        completion(result)
-    }
-}
